High Precision Timer Support
Adds hpt support to several systems.

x_httpd.run_slice( time_usec ) will abort the processing loop when it hits the time limit specified. This allows us to allocate a very fine grained amount of time per X-Plane frame to processing httpd client connections.

x_httpd_request.readClientRequest() is now protected by a HPT that allows a specific max-time window for a legitimate client process to write data to the socket. If no data is rx'd in the specified time limit the connection is dropped with a 500 server error message informing the client of the time out condition.